Avoid 'conversion may change value' cases related to iso files
Avoid some 'conversion from `long long int` to `int` may change value'
and 'conversion from [...] `long int` to `int` may change value' warnings
in iso.cpp, kisofile.cpp and kisodirectory.cpp. Those conversions to `int`
are confirmed because some non-Krusader functions that are used (like
`resize(int)` or `fromTime_t(uint)`) do not accept data types that are
bigger than `int`.

Avoid some 'conversion from `long long int` to `int` may change value'
and 'conversion from [...] `long int` to `int` may change value' warnings
in iso.cpp, kisofile.cpp and kisodirectory.cpp, confirming conversions
to `int` because some non-Krusader functions that are used (like `resize(int)`
or `fromTime_t(uint)`) do not accept data types that are bigger than `int`.

Avoid several 'conversion from [...] `long int` [...]  to `int`' warnings
in some files by means of using a `ssize_t`, `time_t` or `qint64` variables
instead of an `int` one.

Avoid a 'conversion from `int` to `short unsigned int` may change value'
warning in bswap.h by means of using an `(unsigned short)` instead of
a `static_cast<unsigned short>` because, when trying to use a
`static_cast<unsigned short>`, the error '`static_cast` undeclared (first
use in this function)' was seen, as it happens when trying to use `static_cast`
with C (not C++) source code.

Avoid a 'conversion from `int` to `short unsigned int` may change value'
warning in kiso.cpp, confirming a conversion to `__u8` because a non-Krusader
structure does not accept data types that are bigger than `short unsigned
int`.

Revision: https://invent.kde.org/utilities/krusader/-/merge_requests/59

+++ iso/iso.cpp
@@ -133,7 +133,7 @@ bool kio_isoProtocol::checkNewFile(QString fullPath, QString & path, int startse
             if (S_ISLNK(statbuf.st_mode)) {
                 char symDest[256];
                 memset(symDest, 0, 256);
-                int endOfName = readlink(QFile::encodeName(tryPath), symDest, 256);
+                ssize_t endOfName = readlink(QFile::encodeName(tryPath), symDest, 256);
                 if (endOfName != -1) {
                     if (QDir(QString::fromLocal8Bit(symDest)).exists())
                         isFile = false;
@@ -374,9 +374,9 @@ void kio_isoProtocol::getFile(const KIsoFile *isoFileEntry, const QString &path)
             ptrblock_bytes = (nblocks + 1) * 4;
             pointer_block = isoFileEntry->dataAt(hdr->header_size << 2, ptrblock_bytes);
             if ((unsigned long)pointer_block.size() == ptrblock_bytes) {
-                inbuf.resize(block_size2);
+                inbuf.resize(static_cast<int>(block_size2));
                 if (inbuf.size()) {
-                    outbuf.resize(block_size);
+                    outbuf.resize(static_cast<int>(block_size));
 
                     if (outbuf.size())
                         pptr = pointer_block.data();
@@ -433,7 +433,7 @@ void kio_isoProtocol::getFile(const KIsoFile *isoFileEntry, const QString &path)
             if (bytes > fullsize)
                 bytes = fullsize;
             fileData = outbuf;
-            fileData.resize(bytes);
+            fileData.resize(static_cast<int>(bytes));
             fullsize -= bytes;
         } else {
             fileData = isoFileEntry->dataAt(pos, 65536);
